+++ datasets/nmg-extract.sql
@@ -11,10 +11,7 @@ CREATE TABLE "NetworkMemberGroup" (
 	PRIMARY KEY (id)
 );
 INSERT INTO "NetworkMemberGroup" VALUES(1,'00e0U000000ZLaUQAW','0DB0U000000CcHEWA0');
-INSERT INTO "NetworkMemberGroup" VALUES(2,'00e0U000000QmjpQAC','0DB0U000000CcHEWA0');
-INSERT INTO "NetworkMemberGroup" VALUES(3,'00e0U000000QmjqQAC','0DB0U000000CcHEWA0');
-INSERT INTO "NetworkMemberGroup" VALUES(4,'00e0U000000QmjrQAC','0DB0U000000CcHEWA0');
-INSERT INTO "NetworkMemberGroup" VALUES(5,'0PS0U000000FAN2WAO','0DB0U000000CcHEWA0');
+INSERT INTO "NetworkMemberGroup" VALUES(2,'0PS0U000000FAN2WAO','0DB0U000000CcHEWA0');
 CREATE TABLE "PermissionSet" (
 	id VARCHAR(255) NOT NULL, 
 	name VARCHAR(255), 
@@ -77,11 +74,11 @@ INSERT INTO "PermissionSet" VALUES('0PS0U000000P9zFWAS','X00e1a000000N1wLAAS');
 INSERT INTO "PermissionSet" VALUES('0PS0U000000P9zGWAS','X00eB0000000cSXXIA2');
 INSERT INTO "PermissionSet" VALUES('0PS0U000000P9zJWAS','X00e1a000000MWaAAAW');
 INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMnWAO','SurveyCreator');
-INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMuWAO','EinsteinAnalyticsUser');
-INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMvWAO','EinsteinAnalyticsAdmin');
-INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMxWAO','EinsteinAnalyticsPlusAdmin');
-INSERT INTO "PermissionSet" VALUES('0PS0U000000P9zKWAS','SalesConsoleUser');
 INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMwWAO','EinsteinAnalyticsPlusUser');
+INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMxWAO','EinsteinAnalyticsPlusAdmin');
+INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMvWAO','EinsteinAnalyticsAdmin');
+INSERT INTO "PermissionSet" VALUES('0PS0U000000FAMuWAO','EinsteinAnalyticsUser');
+INSERT INTO "PermissionSet" VALUES('0PS0U000000P9zKWAS','SalesConsoleUser');
 CREATE TABLE "Profile" (
 	id VARCHAR(255) NOT NULL, 
 	name VARCHAR(255),
